Minor refactoring segmenting the channel join/leave functionalities from the audio playing/listing functionalities.
