+++ NamazuKingdom/Modules/AudioModule.cs
@@ -22,59 +22,6 @@ namespace NamazuKingdom.Modules
         {
             _lavaNode = node;
         }
-        [Command("join")]
-        public async Task JoinAsync()
-        {
-            if (_lavaNode.HasPlayer(Context.Guild))
-            {
-                await ReplyAsync("I'm already connected to a voice channel!");
-                return;
-            }
-
-            var voiceState = Context.User as IVoiceState;
-            if (voiceState?.VoiceChannel == null)
-            {
-                await ReplyAsync("You must be connected to a voice channel!");
-                return;
-            }
-
-            try
-            {
-                await _lavaNode.JoinAsync(voiceState.VoiceChannel, Context.Channel as ITextChannel);
-                await ReplyAsync($"Joined {voiceState.VoiceChannel.Name}!");
-            }
-            catch (Exception exception)
-            {
-                await ReplyAsync(exception.Message);
-            }
-        }
-
-        [Command("leave")]
-        public async Task LeaveAsync()
-        {
-            if (!_lavaNode.HasPlayer(Context.Guild))
-            {
-                await ReplyAsync("I'm not connected to a voice channel!");
-                return;
-            }
-
-            var voiceState = Context.User as IVoiceState;
-            if (voiceState?.VoiceChannel == null)
-            {
-                await ReplyAsync("You must be connected to a voice channel!");
-                return;
-            }
-
-            try
-            {
-                await _lavaNode.LeaveAsync(voiceState.VoiceChannel);
-                await ReplyAsync($"Left {voiceState.VoiceChannel.Name}!");
-            }
-            catch (Exception exception)
-            {
-                await ReplyAsync(exception.Message);
-            }
-        }
 
         [Command("list_sounds")]
         public async Task ListSoundsAsync()
