+++ NamazuKingdom/Modules/ChannelModule.cs
@@ -1,0 +1,80 @@
+using Discord;
+using Discord.Commands;
+using Discord.WebSocket;
+using Microsoft.Extensions.DependencyInjection;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Xml.Linq;
+using Victoria;
+using Victoria.Enums;
+using Victoria.Responses.Search;
+
+namespace NamazuKingdom.Modules
+{
+
+    public sealed class ChannelModule : ModuleBase<SocketCommandContext>
+    {
+        private readonly LavaNode _lavaNode;
+        public ChannelModule(LavaNode node)
+        {
+            _lavaNode = node;
+        }
+        [Command("join")]
+        public async Task JoinAsync()
+        {
+            if (_lavaNode.HasPlayer(Context.Guild))
+            {
+                await ReplyAsync("I'm already connected to a voice channel!");
+                return;
+            }
+
+            var voiceState = Context.User as IVoiceState;
+            if (voiceState?.VoiceChannel == null)
+            {
+                await ReplyAsync("You must be connected to a voice channel!");
+                return;
+            }
+
+            try
+            {
+                await _lavaNode.JoinAsync(voiceState.VoiceChannel, Context.Channel as ITextChannel);
+                await ReplyAsync($"Joined {voiceState.VoiceChannel.Name}!");
+            }
+            catch (Exception exception)
+            {
+                await ReplyAsync(exception.Message);
+            }
+        }
+
+        [Command("leave")]
+        public async Task LeaveAsync()
+        {
+            if (!_lavaNode.HasPlayer(Context.Guild))
+            {
+                await ReplyAsync("I'm not connected to a voice channel!");
+                return;
+            }
+
+            var voiceState = Context.User as IVoiceState;
+            if (voiceState?.VoiceChannel == null)
+            {
+                await ReplyAsync("You must be connected to a voice channel!");
+                return;
+            }
+
+            try
+            {
+                await _lavaNode.LeaveAsync(voiceState.VoiceChannel);
+                await ReplyAsync($"Left {voiceState.VoiceChannel.Name}!");
+            }
+            catch (Exception exception)
+            {
+                await ReplyAsync(exception.Message);
+            }
+        }
+
+    }
+}
